HW11 BigInteger evaluate

+++ programming/HW11/src/expression/Operation.java
@@ -1,5 +1,7 @@
 package expression;
 
+import java.math.BigInteger;
+
 public abstract class Operation implements CommonExpression {
 
     CommonExpression a;
@@ -18,11 +20,11 @@ public abstract class Operation implements CommonExpression {
 
     @Override
     public int evaluate(int x, int y, int z) {
-        return (int)eval(a.evaluate((long)x,y,z), b.evaluate((long)x,y,z));
+        return (int)eval(a.evaluate(x,y,z), b.evaluate(x,y,z));
     }
 
     @Override
-    public long evaluate(long x, long y, long z) {
+    public BigInteger evaluate(BigInteger x, BigInteger y, BigInteger z) {
         return eval(a.evaluate(x,y,z), b.evaluate(x,y,z));
     }
 
@@ -87,5 +89,5 @@ public abstract class Operation implements CommonExpression {
 
     protected abstract String getSymbol();
     protected abstract int eval(int a, int b);
-    protected abstract long eval(long a, long b);
+    protected abstract BigInteger eval(BigInteger a, BigInteger b);
 }
